Common Words: extract the common words from word files
Fill data directory with word files that conatins hebrew texts and find
the common words. run with -g flag in order to plot a graph in order to find
how many time the most common words appeard in all files,
and run with -t flag in order to output a table contains the frequency of the
most common words.

Signed-off-by: DanielGabay <danielgabay5@gmail.com>