Collect dataset: wordsCutter
This program takes an input (page with handwritten words inside of rectangles)
and split each row of words into folders. Each row of words suppose to be one
of the most common words, that we want to collect data in order to be able to
recognize them in the future.

Signed-off-by: DanielGabay <danielgabay5@gmail.com>